Mailer and get data from api
In this commit I finished the code for the mailing system. The system requires the user to type in their email address, takes it and sends it to the back-end. The node.js mailer sends an email to the  user's email box. I also made a get request and fetched the products list from the server.