Searching system
I made the component, that is supposed to intake user interactions such as searching, adding a product to the list, etc.. I made the searching function as well. It takes the array of all of the products, takes user's query and filters the array accordingly to the user query. The output on the page shows only 5 products.