I made the first prototype of add to cart functionality feature
In this commit I am submitting the first prototype of the add to cart functionality feature. The main code of the function is composed on the ProductSelectionPage(SP).jsx. The function, as well as the variable that holds the added to the card items is then being passed to all of the child components that need access to it.
The function is quite simple. It takes as an argument the product. It checks if the products's id equals to any of the product's ids already in the cart. If yes, it multiplies the quantity of the product with one for every add request made by the user.