I fixed the problem with the variations
I fixed the problem with the variations. I managed to fetch them correctly and to re-render them only when a new product's component is mounted. The variations are fetched directly in the component I use for displaying the products shown in the search bar. It sends a POST request with the product's index to the back-end. Then I map through the array it returns with all of the variations related to this product index and display them in a <select> parsed as <option> items