Worked on relating the products to variations in the shopping cart
I worked on relating the products to the variations in the shopping cart. I assigned a value to each variation item. On change on the select element, it triggers a function that get's the value of the selected variation, and pushes it to the "productWithVar" state variable. Which is then being passed as an argument to the "addProduct" function. I as well modified the the functions in the context, to validate the variations as well