I worked on getting the price of the products
I worked on getting the price for each product from the python calculator. The post request for getting the price is rendered within useEffect in each component, BasketItemsDisplay, that displays a product in the basket in the PageLeftSide component. I had some problems with getting the price correctly and in the correct format. I fixed some of the bugs. However, there are still some bugs in the code that are to be fixed in following commits. I as well fixed some small nonsenses in the code beautified it and commented most of it, so it is easier to read, understand and maintain.