Refactored the code for fetching variations and products
I re-factored the code that fetches the variations and prices for the variations. Now I send one single request and get an array of all the variations. The code is supposed to work faster now. I as well built the calendar element for the contacts page. I am using MUI (material UI) library for the calendar