I worked on the FinalCheck page and the pop-up message there
I worked an almost finished the functionality on the final check page. You can see all of the products in the basked there. You can easily increase or decrease the amount of the products. I worked on the pop-up message. It uses a simple concept. A state variable is assigned with a default value of false. When the variable is false, the window is hidden, when the variable is true the window is hidden. The variable is passed to the pop-up message via props. A prop children is also passed, for the closing button. I did some basic styling as well in order to give my creations some form.