Fixed some issues and cleaned the code
I fixed some small issues, removed some of the unused variables, and re-factored the overall code a little. Added the receipt component in the kit create route. Fixed a little styling as well.  There is an odd bug to be fixed with the pop-up message on the final check page