On of the last commits before testing, fixed a lot of small stuff left for the last moment
Fixed some small bugs the application had. Finished the create kit route as well. Created a new page for contact info, to be applied for the create kit route. However I need to remove it later. Made the logic for the amount picker, now it only shows when on buy kit route. Made the logic for the page left side so i can reuse it. It updates according to which route you take. otherwise I use the same component. Developed the micro copy component that shows on which process step you are on.